Refactor: 조회리스트 생성 로직 DocListAssembler로 분리
